Rust: separate out module level macros
