+++ lib/xdrgen/generators/rust/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "rs-xdr"
+name = "rs-xdr-types"
 version = "0.0.0"
 edition = "2021"
