+++ lib/xdrgen/generators/rust/src/lib.rs
@@ -1,678 +1,5 @@
 #![forbid(unsafe_code)]
 #![cfg_attr(not(feature = "std"), no_std)]
-#![allow(clippy::missing_errors_doc)]
 
-use core::{fmt, fmt::Debug, slice::Iter};
-
-// When feature alloc is turned off use static lifetime Box and Vec types.
-#[cfg(not(feature = "alloc"))]
-mod noalloc {
-    pub mod boxed {
-        pub type Box<T> = &'static T;
-    }
-    pub mod vec {
-        pub type Vec<T> = &'static [T];
-    }
-}
-#[cfg(not(feature = "alloc"))]
-use noalloc::{boxed::Box, vec::Vec};
-
-// When feature std is turned off, but feature alloc is turned on import the
-// alloc crate and use its Box and Vec types.
-#[cfg(all(not(feature = "std"), feature = "alloc"))]
-extern crate alloc;
-#[cfg(all(not(feature = "std"), feature = "alloc"))]
-use alloc::{boxed::Box, vec::Vec};
-
-// TODO: Add support for read/write xdr fns when std not available.
-
-#[cfg(feature = "std")]
-use std::{
-    error, io,
-    io::{Cursor, Read, Write},
-};
-
-#[derive(Debug)]
-pub enum Error {
-    Invalid,
-    LengthExceedsMax,
-    LengthMismatch,
-    NonZeroPadding,
-    #[cfg(feature = "std")]
-    IO(io::Error),
-}
-
-#[cfg(feature = "std")]
-impl error::Error for Error {
-    #[must_use]
-    fn source(&self) -> Option<&(dyn error::Error + 'static)> {
-        match self {
-            Self::IO(e) => Some(e),
-            _ => None,
-        }
-    }
-}
-
-impl fmt::Display for Error {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self {
-            Error::Invalid => write!(f, "xdr value invalid"),
-            Error::LengthExceedsMax => write!(f, "xdr value max length exceeded"),
-            Error::LengthMismatch => write!(f, "xdr value length does not match"),
-            Error::NonZeroPadding => write!(f, "xdr padding contains non-zero bytes"),
-            #[cfg(feature = "std")]
-            Error::IO(e) => write!(f, "{}", e),
-        }
-    }
-}
-
-#[cfg(feature = "std")]
-impl From<io::Error> for Error {
-    #[must_use]
-    fn from(e: io::Error) -> Self {
-        Error::IO(e)
-    }
-}
-
-impl From<Error> for () {
-    fn from(_: Error) {}
-}
-
-#[allow(dead_code)]
-type Result<T> = core::result::Result<T, Error>;
-
-pub trait ReadXdr
-where
-    Self: Sized,
-{
-    #[cfg(feature = "std")]
-    fn read_xdr(r: &mut impl Read) -> Result<Self>;
-
-    #[cfg(feature = "std")]
-    fn read_xdr_into(&mut self, r: &mut impl Read) -> Result<()> {
-        *self = Self::read_xdr(r)?;
-        Ok(())
-    }
-
-    #[cfg(feature = "std")]
-    fn from_xdr_base64(b64: String) -> Result<Self> {
-        let mut b64_reader = Cursor::new(b64);
-        let mut dec = base64::read::DecoderReader::new(&mut b64_reader, base64::STANDARD);
-        let t = Self::read_xdr(&mut dec)?;
-        Ok(t)
-    }
-}
-
-pub trait WriteXdr {
-    #[cfg(feature = "std")]
-    fn write_xdr(&self, w: &mut impl Write) -> Result<()>;
-
-    #[cfg(feature = "std")]
-    fn to_xdr_base64(&self) -> Result<String> {
-        let mut enc = base64::write::EncoderStringWriter::new(base64::STANDARD);
-        self.write_xdr(&mut enc)?;
-        let b64 = enc.into_inner();
-        Ok(b64)
-    }
-}
-
-/// `Pad_len` returns the number of bytes to pad an XDR value of the given
-/// length to make the final serialized size a multiple of 4.
-#[cfg(feature = "std")]
-fn pad_len(len: usize) -> usize {
-    (4 - (len % 4)) % 4
-}
-
-impl ReadXdr for i32 {
-    #[cfg(feature = "std")]
-    fn read_xdr(r: &mut impl Read) -> Result<Self> {
-        let mut b = [0u8; 4];
-        r.read_exact(&mut b)?;
-        let i = i32::from_be_bytes(b);
-        Ok(i)
-    }
-}
-
-impl WriteXdr for i32 {
-    #[cfg(feature = "std")]
-    fn write_xdr(&self, w: &mut impl Write) -> Result<()> {
-        let b: [u8; 4] = self.to_be_bytes();
-        w.write_all(&b)?;
-        Ok(())
-    }
-}
-
-impl ReadXdr for u32 {
-    #[cfg(feature = "std")]
-    fn read_xdr(r: &mut impl Read) -> Result<Self> {
-        let mut b = [0u8; 4];
-        r.read_exact(&mut b)?;
-        let i = u32::from_be_bytes(b);
-        Ok(i)
-    }
-}
-
-impl WriteXdr for u32 {
-    #[cfg(feature = "std")]
-    fn write_xdr(&self, w: &mut impl Write) -> Result<()> {
-        let b: [u8; 4] = self.to_be_bytes();
-        w.write_all(&b)?;
-        Ok(())
-    }
-}
-
-impl ReadXdr for i64 {
-    #[cfg(feature = "std")]
-    fn read_xdr(r: &mut impl Read) -> Result<Self> {
-        let mut b = [0u8; 8];
-        r.read_exact(&mut b)?;
-        let i = i64::from_be_bytes(b);
-        Ok(i)
-    }
-}
-
-impl WriteXdr for i64 {
-    #[cfg(feature = "std")]
-    fn write_xdr(&self, w: &mut impl Write) -> Result<()> {
-        let b: [u8; 8] = self.to_be_bytes();
-        w.write_all(&b)?;
-        Ok(())
-    }
-}
-
-impl ReadXdr for u64 {
-    #[cfg(feature = "std")]
-    fn read_xdr(r: &mut impl Read) -> Result<Self> {
-        let mut b = [0u8; 8];
-        r.read_exact(&mut b)?;
-        let i = u64::from_be_bytes(b);
-        Ok(i)
-    }
-}
-
-impl WriteXdr for u64 {
-    #[cfg(feature = "std")]
-    fn write_xdr(&self, w: &mut impl Write) -> Result<()> {
-        let b: [u8; 8] = self.to_be_bytes();
-        w.write_all(&b)?;
-        Ok(())
-    }
-}
-
-impl ReadXdr for f32 {
-    #[cfg(feature = "std")]
-    fn read_xdr(_r: &mut impl Read) -> Result<Self> {
-        todo!()
-    }
-}
-
-impl WriteXdr for f32 {
-    #[cfg(feature = "std")]
-    fn write_xdr(&self, _w: &mut impl Write) -> Result<()> {
-        todo!()
-    }
-}
-
-impl ReadXdr for f64 {
-    #[cfg(feature = "std")]
-    fn read_xdr(_r: &mut impl Read) -> Result<Self> {
-        todo!()
-    }
-}
-
-impl WriteXdr for f64 {
-    #[cfg(feature = "std")]
-    fn write_xdr(&self, _w: &mut impl Write) -> Result<()> {
-        todo!()
-    }
-}
-
-impl ReadXdr for bool {
-    #[cfg(feature = "std")]
-    fn read_xdr(r: &mut impl Read) -> Result<Self> {
-        let i = u32::read_xdr(r)?;
-        let b = i == 1;
-        Ok(b)
-    }
-}
-
-impl WriteXdr for bool {
-    #[cfg(feature = "std")]
-    fn write_xdr(&self, w: &mut impl Write) -> Result<()> {
-        let i: u32 = if *self { 1 } else { 0 };
-        i.write_xdr(w)?;
-        Ok(())
-    }
-}
-
-impl<T: ReadXdr> ReadXdr for Option<T> {
-    #[cfg(feature = "std")]
-    fn read_xdr(r: &mut impl Read) -> Result<Self> {
-        let i = u32::read_xdr(r)?;
-        match i {
-            0 => Ok(None),
-            1 => {
-                let t = T::read_xdr(r)?;
-                Ok(Some(t))
-            }
-            _ => Err(Error::Invalid),
-        }
-    }
-}
-
-impl<T: WriteXdr> WriteXdr for Option<T> {
-    #[cfg(feature = "std")]
-    fn write_xdr(&self, w: &mut impl Write) -> Result<()> {
-        if let Some(t) = self {
-            1u32.write_xdr(w)?;
-            t.write_xdr(w)?;
-        } else {
-            0u32.write_xdr(w)?;
-        }
-        Ok(())
-    }
-}
-
-impl<T: ReadXdr> ReadXdr for Box<T> {
-    #[cfg(feature = "std")]
-    fn read_xdr(r: &mut impl Read) -> Result<Self> {
-        let t = T::read_xdr(r)?;
-        Ok(Box::new(t))
-    }
-}
-
-impl<T: WriteXdr> WriteXdr for Box<T> {
-    #[cfg(feature = "std")]
-    fn write_xdr(&self, w: &mut impl Write) -> Result<()> {
-        T::write_xdr(self, w)?;
-        Ok(())
-    }
-}
-
-impl ReadXdr for () {
-    #[cfg(feature = "std")]
-    fn read_xdr(_r: &mut impl Read) -> Result<Self> {
-        Ok(())
-    }
-}
-
-impl WriteXdr for () {
-    #[cfg(feature = "std")]
-    fn write_xdr(&self, _w: &mut impl Write) -> Result<()> {
-        Ok(())
-    }
-}
-
-impl<const N: usize> ReadXdr for [u8; N] {
-    #[cfg(feature = "std")]
-    fn read_xdr(r: &mut impl Read) -> Result<Self> {
-        let mut arr = [0u8; N];
-        r.read_exact(&mut arr)?;
-
-        let pad = &mut [0u8; 3][..pad_len(N)];
-        r.read_exact(pad)?;
-        if pad.iter().any(|b| *b != 0) {
-            return Err(Error::NonZeroPadding);
-        }
-
-        Ok(arr)
-    }
-}
-
-impl<const N: usize> WriteXdr for [u8; N] {
-    #[cfg(feature = "std")]
-    fn write_xdr(&self, w: &mut impl Write) -> Result<()> {
-        w.write_all(self)?;
-        w.write_all(&[0u8; 3][..pad_len(N)])?;
-        Ok(())
-    }
-}
-
-impl<T: ReadXdr, const N: usize> ReadXdr for [T; N] {
-    #[cfg(feature = "std")]
-    fn read_xdr(r: &mut impl Read) -> Result<Self> {
-        let mut vec = Vec::with_capacity(N);
-        for _ in 0..N {
-            let t = T::read_xdr(r)?;
-            vec.push(t);
-        }
-        let arr: [T; N] = vec.try_into().unwrap_or_else(|_: Vec<T>| unreachable!());
-        Ok(arr)
-    }
-}
-
-impl<T: WriteXdr, const N: usize> WriteXdr for [T; N] {
-    #[cfg(feature = "std")]
-    fn write_xdr(&self, w: &mut impl Write) -> Result<()> {
-        for t in self {
-            t.write_xdr(w)?;
-        }
-        Ok(())
-    }
-}
-
-#[cfg(feature = "alloc")]
-#[derive(Debug, Clone, Hash, PartialEq, Eq, PartialOrd, Ord)]
-pub struct VecM<T, const MAX: u32 = { u32::MAX }>(Vec<T>);
-
-#[cfg(not(feature = "alloc"))]
-#[derive(Debug, Clone, Hash, PartialEq, Eq, PartialOrd, Ord)]
-pub struct VecM<T, const MAX: u32 = { u32::MAX }>(Vec<T>)
-where
-    T: 'static;
-
-impl<T, const MAX: u32> VecM<T, MAX> {
-    #[must_use]
-    pub fn len(&self) -> usize {
-        self.0.len()
-    }
-
-    #[must_use]
-    pub fn is_empty(&self) -> bool {
-        self.0.is_empty()
-    }
-
-    #[must_use]
-    pub fn to_vec(self) -> Vec<T> {
-        self.into()
-    }
-
-    #[must_use]
-    pub fn as_vec(&self) -> &Vec<T> {
-        self.as_ref()
-    }
-
-    #[must_use]
-    pub fn as_slice(&self) -> &[T] {
-        self.as_ref()
-    }
-
-    pub fn iter(&self) -> Iter<'_, T> {
-        self.0.iter()
-    }
-}
-
-impl<T, const MAX: u32> TryFrom<Vec<T>> for VecM<T, MAX> {
-    type Error = Error;
-
-    fn try_from(v: Vec<T>) -> Result<Self> {
-        let len: u32 = v.len().try_into().map_err(|_| Error::LengthExceedsMax)?;
-        if len <= MAX {
-            Ok(VecM(v))
-        } else {
-            Err(Error::LengthExceedsMax)
-        }
-    }
-}
-
-impl<T, const MAX: u32> From<VecM<T, MAX>> for Vec<T> {
-    #[must_use]
-    fn from(v: VecM<T, MAX>) -> Self {
-        v.0
-    }
-}
-
-impl<T, const MAX: u32> AsRef<Vec<T>> for VecM<T, MAX> {
-    #[must_use]
-    fn as_ref(&self) -> &Vec<T> {
-        &self.0
-    }
-}
-
-#[cfg(feature = "alloc")]
-impl<T: Clone, const MAX: u32> TryFrom<&[T]> for VecM<T, MAX> {
-    type Error = Error;
-
-    fn try_from(v: &[T]) -> Result<Self> {
-        let len: u32 = v.len().try_into().map_err(|_| Error::LengthExceedsMax)?;
-        if len <= MAX {
-            Ok(VecM(v.to_vec()))
-        } else {
-            Err(Error::LengthExceedsMax)
-        }
-    }
-}
-
-impl<T, const MAX: u32> AsRef<[T]> for VecM<T, MAX> {
-    #[cfg(feature = "alloc")]
-    #[must_use]
-    fn as_ref(&self) -> &[T] {
-        self.0.as_ref()
-    }
-    #[cfg(not(feature = "alloc"))]
-    #[must_use]
-    fn as_ref(&self) -> &[T] {
-        self.0
-    }
-}
-
-#[cfg(feature = "alloc")]
-impl<T: Clone, const N: usize, const MAX: u32> TryFrom<[T; N]> for VecM<T, MAX> {
-    type Error = Error;
-
-    fn try_from(v: [T; N]) -> Result<Self> {
-        let len: u32 = v.len().try_into().map_err(|_| Error::LengthExceedsMax)?;
-        if len <= MAX {
-            Ok(VecM(v.to_vec()))
-        } else {
-            Err(Error::LengthExceedsMax)
-        }
-    }
-}
-
-#[cfg(feature = "alloc")]
-impl<T: Clone, const N: usize, const MAX: u32> TryFrom<VecM<T, MAX>> for [T; N] {
-    type Error = VecM<T, MAX>;
-
-    fn try_from(v: VecM<T, MAX>) -> core::result::Result<Self, Self::Error> {
-        let s: [T; N] = v.0.try_into().map_err(|v: Vec<T>| VecM::<T, MAX>(v))?;
-        Ok(s)
-    }
-}
-
-#[cfg(feature = "alloc")]
-impl<T: Clone, const N: usize, const MAX: u32> TryFrom<&[T; N]> for VecM<T, MAX> {
-    type Error = Error;
-
-    fn try_from(v: &[T; N]) -> Result<Self> {
-        let len: u32 = v.len().try_into().map_err(|_| Error::LengthExceedsMax)?;
-        if len <= MAX {
-            Ok(VecM(v.to_vec()))
-        } else {
-            Err(Error::LengthExceedsMax)
-        }
-    }
-}
-
-#[cfg(not(feature = "alloc"))]
-impl<T: Clone, const N: usize, const MAX: u32> TryFrom<&'static [T; N]> for VecM<T, MAX>
-where
-    T: 'static,
-{
-    type Error = Error;
-
-    fn try_from(v: &'static [T; N]) -> Result<Self> {
-        let len: u32 = v.len().try_into().map_err(|_| Error::LengthExceedsMax)?;
-        if len <= MAX {
-            Ok(VecM(v))
-        } else {
-            Err(Error::LengthExceedsMax)
-        }
-    }
-}
-
-impl<const MAX: u32> ReadXdr for VecM<u8, MAX> {
-    #[cfg(feature = "std")]
-    fn read_xdr(r: &mut impl Read) -> Result<Self> {
-        let len: u32 = u32::read_xdr(r)?;
-        if len > MAX {
-            return Err(Error::LengthExceedsMax);
-        }
-
-        let mut vec = vec![0u8; len as usize];
-        r.read_exact(&mut vec)?;
-
-        let pad = &mut [0u8; 3][..pad_len(len as usize)];
-        r.read_exact(pad)?;
-        if pad.iter().any(|b| *b != 0) {
-            return Err(Error::NonZeroPadding);
-        }
-
-        Ok(VecM(vec))
-    }
-}
-
-impl<const MAX: u32> WriteXdr for VecM<u8, MAX> {
-    #[cfg(feature = "std")]
-    fn write_xdr(&self, w: &mut impl Write) -> Result<()> {
-        let len: u32 = self.len().try_into().map_err(|_| Error::LengthExceedsMax)?;
-        len.write_xdr(w)?;
-
-        w.write_all(&self.0)?;
-
-        w.write_all(&[0u8; 3][..pad_len(len as usize)])?;
-
-        Ok(())
-    }
-}
-
-impl<T: ReadXdr, const MAX: u32> ReadXdr for VecM<T, MAX> {
-    #[cfg(feature = "std")]
-    fn read_xdr(r: &mut impl Read) -> Result<Self> {
-        let len = u32::read_xdr(r)?;
-        if len > MAX {
-            return Err(Error::LengthExceedsMax);
-        }
-
-        let mut vec = Vec::with_capacity(len as usize);
-        for _ in 0..len {
-            let t = T::read_xdr(r)?;
-            vec.push(t);
-        }
-
-        Ok(VecM(vec))
-    }
-}
-
-impl<T: WriteXdr, const MAX: u32> WriteXdr for VecM<T, MAX> {
-    #[cfg(feature = "std")]
-    fn write_xdr(&self, w: &mut impl Write) -> Result<()> {
-        let len: u32 = self.len().try_into().map_err(|_| Error::LengthExceedsMax)?;
-        len.write_xdr(w)?;
-
-        for t in &self.0 {
-            t.write_xdr(w)?;
-        }
-
-        Ok(())
-    }
-}
-
-#[cfg(all(test, feature = "std"))]
-mod tests {
-    use std::io::Cursor;
-
-    use crate::WriteXdr;
-
-    use super::{Error, ReadXdr, VecM};
-
-    #[test]
-    pub fn vec_u8_read_without_padding() {
-        let mut buf = Cursor::new(vec![0, 0, 0, 4, 2, 2, 2, 2]);
-        let v = VecM::<u8, 8>::read_xdr(&mut buf).unwrap();
-        assert_eq!(v.to_vec(), vec![2, 2, 2, 2]);
-    }
-
-    #[test]
-    pub fn vec_u8_read_with_padding() {
-        let mut buf = Cursor::new(vec![0, 0, 0, 1, 2, 0, 0, 0]);
-        let v = VecM::<u8, 8>::read_xdr(&mut buf).unwrap();
-        assert_eq!(v.to_vec(), vec![2]);
-    }
-
-    #[test]
-    pub fn vec_u8_read_with_insufficient_padding() {
-        let mut buf = Cursor::new(vec![0, 0, 0, 1, 2, 0, 0]);
-        let res = VecM::<u8, 8>::read_xdr(&mut buf);
-        match res {
-            Err(Error::IO(_)) => (),
-            _ => panic!("expected IO error got {:?}", res),
-        }
-    }
-
-    #[test]
-    pub fn vec_u8_read_with_non_zero_padding() {
-        let mut buf = Cursor::new(vec![0, 0, 0, 1, 2, 3, 0, 0]);
-        let res = VecM::<u8, 8>::read_xdr(&mut buf);
-        match res {
-            Err(Error::NonZeroPadding) => (),
-            _ => panic!("expected NonZeroPadding got {:?}", res),
-        }
-    }
-
-    #[test]
-    pub fn vec_u8_write_without_padding() {
-        let mut buf = vec![];
-        let v: VecM<u8, 8> = vec![2, 2, 2, 2].try_into().unwrap();
-        v.write_xdr(&mut Cursor::new(&mut buf)).unwrap();
-        assert_eq!(buf, vec![0, 0, 0, 4, 2, 2, 2, 2]);
-    }
-
-    #[test]
-    pub fn vec_u8_write_with_padding() {
-        let mut buf = vec![];
-        let v: VecM<u8, 8> = vec![2].try_into().unwrap();
-        v.write_xdr(&mut Cursor::new(&mut buf)).unwrap();
-        assert_eq!(buf, vec![0, 0, 0, 1, 2, 0, 0, 0]);
-    }
-
-    #[test]
-    pub fn arr_u8_read_without_padding() {
-        let mut buf = Cursor::new(vec![2, 2, 2, 2]);
-        let v = <[u8; 4]>::read_xdr(&mut buf).unwrap();
-        assert_eq!(v, [2, 2, 2, 2]);
-    }
-
-    #[test]
-    pub fn arr_u8_read_with_padding() {
-        let mut buf = Cursor::new(vec![2, 0, 0, 0]);
-        let v = <[u8; 1]>::read_xdr(&mut buf).unwrap();
-        assert_eq!(v, [2]);
-    }
-
-    #[test]
-    pub fn arr_u8_read_with_insufficient_padding() {
-        let mut buf = Cursor::new(vec![2, 0, 0]);
-        let res = <[u8; 1]>::read_xdr(&mut buf);
-        match res {
-            Err(Error::IO(_)) => (),
-            _ => panic!("expected IO error got {:?}", res),
-        }
-    }
-
-    #[test]
-    pub fn arr_u8_read_with_non_zero_padding() {
-        let mut buf = Cursor::new(vec![2, 3, 0, 0]);
-        let res = <[u8; 1]>::read_xdr(&mut buf);
-        match res {
-            Err(Error::NonZeroPadding) => (),
-            _ => panic!("expected NonZeroPadding got {:?}", res),
-        }
-    }
-
-    #[test]
-    pub fn arr_u8_write_without_padding() {
-        let mut buf = vec![];
-        [2u8, 2, 2, 2]
-            .write_xdr(&mut Cursor::new(&mut buf))
-            .unwrap();
-        assert_eq!(buf, vec![2, 2, 2, 2]);
-    }
-
-    #[test]
-    pub fn arr_u8_write_with_padding() {
-        let mut buf = vec![];
-        [2u8].write_xdr(&mut Cursor::new(&mut buf)).unwrap();
-        assert_eq!(buf, vec![2, 0, 0, 0]);
-    }
-}
+mod types;
+pub use types::*;
